Two functions which store a matrix and cache its inverse.
The first function creates a vector which is a list which contains four
functions which set the value of the matrix, retrive the value of the
matrix, set the value of the inverse of the matrix, and gets the value
of the inverse of the matrix.

The second function computes the inverse of a matrix, but first checks
to see if the inverse has already been calculated, and if so, retrieves
it.